+++ Source/Core/EllipsoidGeometry.js
@@ -224,7 +224,7 @@ define([
             }
 
             attributes.position = new GeometryAttribute({
-                componentDatatype : ComponentDatatype.FLOAT,
+                componentDatatype : ComponentDatatype.DOUBLE,
                 componentsPerAttribute : 3,
                 values : flattenedPositions
             });
